fix: remove unnecessary type conversions

+++ x-pack/filebeat/input/netflow/definitions.go
@@ -169,7 +169,7 @@ func loadFields(def map[interface{}]interface{}, pem uint32, dest fields.FieldDi
 			return fmt.Errorf("bad field ID %d: should have two items (type, name) or one (:skip) (Got %+v)", fieldID, list)
 		}
 		key := fields.Key{
-			EnterpriseID: uint32(pem),
+			EnterpriseID: pem,
 			FieldID:      uint16(fieldID),
 		}
 		if _, exists := dest[key]; exists {
